requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
openai>=1.58.0
anthropic>=0.40.0
aiohttp>=3.11.0
python-multipart>=0.0.20
redis>=5.2.0
